Rearrange code, add utils with function to automatically covert swagger yaml to json format
